Show machine stop event details on machine status
