+++ internal/command/machine/status.go
@@ -103,14 +103,22 @@ func runMachineStatus(ctx context.Context) (err error) {
 
 	for _, event := range machine.Events {
 		timeInUTC := time.Unix(0, event.Timestamp*int64(time.Millisecond))
-		eventLogs = append(eventLogs, []string{
+		fields := []string{
 			event.Status,
 			event.Type,
 			event.Source,
 			timeInUTC.Format(time.RFC3339Nano),
-		})
+		}
+
+		if event.Request != nil && event.Request.ExitEvent != nil {
+			exitEvent := event.Request.ExitEvent
+			fields = append(fields, fmt.Sprintf("exit_code=%d,oom_killed=%t,requested_stop=%t",
+				exitEvent.GuestExitCode, exitEvent.OOMKilled, exitEvent.RequestedStop))
+		}
+
+		eventLogs = append(eventLogs, fields)
 	}
-	_ = render.Table(io.Out, "Event Logs", eventLogs, "State", "Event", "Source", "Timestamp")
+	_ = render.Table(io.Out, "Event Logs", eventLogs, "State", "Event", "Source", "Timestamp", "Info")
 
 	if flag.GetBool(ctx, "display-config") {
 		var prettyConfig []byte
